requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
dj-database-url>=3.0.1
Pillow>=10.2.0
numpy>=1.24.3
scipy>=1.10.1
matplotlib>=3.7.1
seaborn>=0.12.2
pandas>=2.0.2
h5py>=3.8.0
netCDF4>=1.6.5
xarray>=2023.5.0
cartopy>=0.21.1
geopandas>=0.12.2
folium>=0.14.0
plotly>=5.14.1
dash>=2.9.3
dash-bootstrap-components>=1.4.1
gunicorn>=21.2.0
whitenoise>=6.5.0
python-decouple>=3.8
google-cloud-storage>=2.10.0 